refactor: refactor the mklittlefs for binary consistency
To guarantee the binary consistency, following steps are used:

1. Remove the timestamp
Don't add timestamp of files and root. The origin tools use LFS custom
attributes to add timestamp of files and root. Removing it, the image
will be time-independent.

2. Sort the directory entry before adding them to the image
As the readdir does not guarantee an order, the creating sequence of
filesystem image is not guaranted. Here, we use BFS instead of DFS to
tranverse the target directory, and sort the directory entries before
adding them to the image.

Signed-off-by: Far <yesiyuan2@huawei.com>
Change-Id: I94adafc758cefc690c934592b4be58f2c987c9e0